+++ tg_searcher/tg_searcher/backend_bot.py
@@ -8,8 +8,10 @@ from telethon.tl.patched import Message as TgMessage
 from telethon.tl.types import User
 
 from .indexer import Indexer, IndexMsg
-from .common import CommonBotConfig, escape_content, get_share_id, get_logger, format_entity_name, brief_content
+from .common import CommonBotConfig, escape_content, get_share_id, get_logger, format_entity_name, brief_content, \
+    EntityNotFoundError
 from .session import ClientSession
+
 
 class BackendBotConfig:
     def __init__(self, **kw):
@@ -17,10 +19,6 @@ class BackendBotConfig:
         self.excluded_chats: Set[int] = set(get_share_id(chat_id)
                                             for chat_id in kw.get('exclude_chats', []))
 
-class EntityNotFoundError(Exception):
-    def __init__(self, entity):
-        super().__init__(f'Cannot find entity of id {entity}')
-        self.entity = entity
 
 class BackendBot:
     def __init__(self, common_cfg: CommonBotConfig, cfg: BackendBotConfig,
@@ -79,7 +77,8 @@ class BackendBot:
                 )
                 self._indexer.add_document(msg, writer)
                 self.newest_msg[share_id] = msg
-                await call_back(tg_message.id)
+                if call_back:
+                    await call_back(tg_message.id)
         writer.commit()
 
     def clear(self, chat_ids: Optional[List[int]] = None):
@@ -96,31 +95,43 @@ class BackendBot:
     async def find_chat_id(self, q: str) -> List[int]:
         return await self.session.find_chat_id(q)
 
-    async def get_index_status(self):
+    async def get_index_status(self, length_limit: int = 4000):
         # TODO: add session and frontend name
+        cur_len = 0
         sb = [  # string builder
             f'后端 "{self.id}"（session: "{self.session.name}"）总消息数: <b>{self._indexer.ix.doc_count()}</b>\n\n'
         ]
+        overflow_msg = f'\n\n由于 Telegram 消息长度限制，部分对话的统计信息没有展示'
+
+        def append_msg(msg_list: List[str]):  # return whether overflow
+            nonlocal cur_len, sb
+            total_len = sum(len(msg) for msg in msg_list)
+            if cur_len + total_len > length_limit - len(overflow_msg):
+                return True
+            else:
+                cur_len += total_len
+                for msg in msg_list:
+                    sb.append(msg)
+                    return False
+
         if self._cfg.monitor_all:
-            sb.append(f'{len(self.excluded_chats)} 个对话被禁止索引\n')
+            append_msg([f'{len(self.excluded_chats)} 个对话被禁止索引\n'])
             for chat_id in self.excluded_chats:
-                sb.append(f'- {await self.format_dialog_html(chat_id)}\n')
+                append_msg([f'- {await self.format_dialog_html(chat_id)}\n'])
             sb.append('\n')
 
-        sb.append(f'总计 {len(self.monitored_chats)} 个对话被加入了索引：\n')
-        print_limit = 100  # since telegram can send at most 4096 chars per msg
+        append_msg([f'总计 {len(self.monitored_chats)} 个对话被加入了索引：\n'])
         for chat_id in self.monitored_chats:
-            if print_limit > 0:
-                print_limit -= 1
-            else:
-                break
+            msg_for_chat = []
             num = self._indexer.count_by_query(chat_id=str(chat_id))
-            sb.append(f'- {await self.format_dialog_html(chat_id)} '
-                      f'共 {num} 条消息\n')
+            msg_for_chat.append(f'- {await self.format_dialog_html(chat_id)} 共 {num} 条消息\n')
             if newest_msg := self.newest_msg.get(chat_id, None):
-                sb.append(f'  最新消息：<a href="{newest_msg.url}">{brief_content(newest_msg.content)}</a>\n')
-        if print_limit == 0:
-            sb.append(f'\n由于 Telegram 的消息长度限制，最多只显示 100 个对话')
+                msg_for_chat.append(f'  最新消息：<a href="{newest_msg.url}">{brief_content(newest_msg.content)}</a>\n')
+            if append_msg(msg_for_chat):
+                # if overflow
+                sb.append(overflow_msg)
+                break
+
         return ''.join(sb)
 
     async def translate_chat_id(self, chat_id: int) -> str:
@@ -130,14 +141,7 @@ class BackendBot:
             return '[无法获取名称]'
 
     async def str_to_chat_id(self, chat: str) -> int:
-        try:
-            return int(chat)
-        except ValueError:
-            try:
-                entity = await self.session.get_entity(chat)
-            except ValueError:
-                raise EntityNotFoundError(chat)
-            return get_share_id(entity.id)
+        return await self.session.str_to_chat_id(chat)
 
     async def format_dialog_html(self, chat_id: int):
         # TODO: handle PM URL